Update to attribute table classes to provide importer from hdf5 file format. All commands have been switched over to use the hdf5 exporter. This is because the hdf5 files are small and also significantly faster to import and export.
All commands have be changed to have either an ascii or hdf5 file inputted and they will automatically decide on the importer to use (tests whether file path is a valid hdf5 file if not then assumed to be an ascii file).

Finally, a command to export an attribute table to original ascii format has been provided - might provide easier transfer to other software packages (e.g., R or python).