Forgot to increment the CMakeLists.txt svn version number.
Update which includes a working version of the triangulation based warping using CGAL for the warping. Note this makes CGAL a dependency of the library (although it was before as it was trying to link but the code was commented out so the headers didn't need to be in the path).