Added a function to calculate a segmentation 'quality' metric for a given a segmentation and utility functions (segutils) which will loop through a range of clumps or minimum pixel sizes and produce a summary stats file for the segmentation parameters. These can be used to identify the 'best' segmentation parameters to be used.
The method of Johnson & Xie (2011) is the one that has been implemented. I would recommend using any of these methods with caution as the result many still not be appropriate for your application, however it maybe be useful and does allow a comparison to be made.

Johnson, B., & Xie, Z. (2011). Unsupervised image segmentation evaluation and refinement using a multi-scale approach. ISPRS Journal of Photogrammetry and Remote Sensing, 66(4), 473?483. doi:10.1016/j.isprsjprs.2011.02.006